pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-async"
dynamic = ["version"]
description = "Interactive Brokers Async Trading Framework for Python 3.12+"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Kelvin Gao", email = "89156201@qq.com"}
]
maintainers = [
    {name = "Kelvin Gao", email = "89156201@qq.com"}
]
keywords = ["trading", "interactive-brokers", "async", "finance", "algorithmic-trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.12"
dependencies = [
    "aiozmq>=1.0.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "bump2version>=1.0.1",
    "coverage>=7.10.2",
    # "ezib-async",  # TODO: Replace with published version when available
    "ib-async>=0.9.86", # Alternative IB async library
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "msgspec>=0.19.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
    "uvicorn>=0.30.0",
    "twine>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/kelvingao/quant_async"
Repository = "https://github.com/kelvingao/quant_async.git"
Documentation = "https://github.com/kelvingao/quant_async#readme"
Issues = "https://github.com/kelvingao/quant_async/issues"

[tool.setuptools]
packages = ["quant_async"]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "quant_async.version.__version__"}

# [tool.uv.sources]
# ezib-async = { path = "../ezib_async", editable = true }  # Commented out for PyPI

[dependency-groups]
dev = [
    "alembic>=1.15.2",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.7",
]
